Modified different aspects of the GA
Notably the genome is grouped into "chromosomes" where each chromosome consists of all the parameters that belong to a convolutional or dense layer.
An offspring is created by mixing the chromosomes of the parents as well as randomly changing some parameters by picking a value from the initial distribution of parameters.
There is now also a hyperparameter to control for how many new children are added to the next generation. The children replace the worst performing individuals.